Small change to address the issue when your list of edges has only 1 entry
I just ran into an issue where the fortify was failing when the edge list of one row. I'm not sure this is the best way to handle the issue but I wanted to provide a possible solution. the  issue is that  `edges[, 1:2]` will return a vector instead of a data frame which will cause `nrow` to throw an error.